tests/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.3.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.3.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pytest==8.3.5
pytest-asyncio==0.25.3
python-dotenv==1.0.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
yarl==1.18.3
Werkzeug==3.1.3
